Release 1.4.2 (#2852)
* Adding react-addons-shallow-compare dependency (#2829)

* Adding react-addons-shallow-compare dependencie


* Update package.json

* Stripe patches

- Add handling for Stripe API Key handling @lcampanis

* Meteor 1.5.2

- Update to Meteor 1.5.2

* Removed unused import

* Mute dependencies

- will be updated in next release

* Remove npm-shrinkwrap

* Ignore prop type warnings

- add TODO

* Skip Stripe tests

* Skip Stripe tests
