+++ .meteor/packages
@@ -11,13 +11,13 @@ meteor-base@1.1.0             # Packages every Meteor app needs to have
 mobile-experience@1.0.4       # Packages for a great mobile UX
 blaze-html-templates@1.0.4    # Compile .html files into Meteor Blaze views
 es5-shim@4.6.15               # ECMAScript 5 compatibility for older browsers.
-ecmascript@0.8.1              # Enable ECMAScript2015+ syntax in app code
+ecmascript@0.8.2              # Enable ECMAScript2015+ syntax in app code
 audit-argument-checks@1.0.7   # ensure meteor method argument validation
 browser-policy@1.1.0          # security-related policies enforced by newer browsers
 juliancwirko:postcss          # CSS post-processing plugin (replaces standard-minifier-css)
 session@1.1.7                 # ReactiveDict whose contents are preserved across Hot Code Push
 tracker@1.1.3                 # Meteor transparent reactive programming library
-mongo@1.1.19
+mongo@1.2.0
 random@1.0.10
 reactive-var@1.0.11
 reactive-dict@1.1.9
@@ -27,7 +27,7 @@ ddp-rate-limiter@1.0.7
 underscore@1.0.10
 logging@1.1.17
 reload@1.1.11
-ejson@1.0.13
+ejson@1.0.14
 less@2.7.9
 service-configuration@1.0.11
 mdg:validated-method
@@ -35,7 +35,7 @@ shell-server@0.2.4
 dynamic-import@0.1.1
 
 # Meteor Auth Packages
-accounts-base@1.3.1
+accounts-base@1.3.2
 accounts-password@1.4.0
 accounts-facebook@1.2.1
 accounts-google@1.2.0
