+++ .meteor/release
@@ -1,1 +1,1 @@
-METEOR@1.5.1
+METEOR@1.5.2
