+++ .meteor/versions
@@ -1,6 +1,6 @@
 abernix:minifier-js@2.1.0
 abernix:standard-minifier-js@2.1.0
-accounts-base@1.3.1
+accounts-base@1.3.2
 accounts-facebook@1.2.1
 accounts-google@1.2.0
 accounts-oauth@1.1.15
@@ -18,14 +18,14 @@ aldeed:template-extension@4.1.0
 allow-deny@1.0.6
 audit-argument-checks@1.0.7
 autoupdate@1.3.12
-babel-compiler@6.19.4
+babel-compiler@6.20.0
 babel-runtime@1.0.1
 base64@1.0.10
 binary-heap@1.0.10
 blaze@2.3.2
 blaze-html-templates@1.1.2
 blaze-tools@1.0.10
-boilerplate-generator@1.1.1
+boilerplate-generator@1.2.0
 bozhao:accounts-instagram@0.2.2
 browser-policy@1.1.0
 browser-policy-common@1.0.11
@@ -55,10 +55,11 @@ cfs:ui@0.1.3
 cfs:upload-http@0.0.20
 cfs:worker@0.1.4
 check@1.2.5
-coffeescript@1.12.6_1
+coffeescript@1.12.7_1
+coffeescript-compiler@1.12.7_1
 dburles:factory@1.1.0
 ddp@1.3.0
-ddp-client@2.0.0
+ddp-client@2.1.0
 ddp-common@1.2.9
 ddp-rate-limiter@1.0.7
 ddp-server@2.0.0
@@ -71,7 +72,7 @@ ecmascript@0.8.2
 ecmascript-runtime@0.4.1
 ecmascript-runtime-client@0.4.3
 ecmascript-runtime-server@0.4.1
-ejson@1.0.13
+ejson@1.0.14
 email@1.2.3
 es5-shim@4.6.15
 facebook-config-ui@1.0.0
@@ -103,18 +104,19 @@ logging@1.1.17
 matb33:collection-hooks@0.8.4
 mdg:validated-method@1.1.0
 mdg:validation-error@0.5.1
-meteor@1.7.0
+meteor@1.7.1
 meteor-base@1.1.0
 meteorhacks:ssr@2.2.0
 meteorhacks:subs-manager@1.6.4
 minifier-css@1.2.16
-minimongo@1.2.1
+minimongo@1.3.0
 mobile-experience@1.0.4
 mobile-status-bar@1.0.14
-modules@0.9.2
+modules@0.10.0
 modules-runtime@0.8.0
 momentjs:moment@2.18.1
-mongo@1.1.22
+mongo@1.2.0
+mongo-dev-server@1.0.1
 mongo-id@1.0.6
 mongo-livedata@1.0.12
 mrt:later@1.6.1
@@ -132,7 +134,7 @@ percolate:migrations@0.9.8
 practicalmeteor:chai@2.1.0_1
 practicalmeteor:mocha-core@1.0.1
 practicalmeteor:sinon@1.14.1_2
-promise@0.8.9
+promise@0.9.0
 raix:eventemitter@0.1.3
 raix:ui-dropped-event@0.0.7
 random@1.0.10
@@ -148,7 +150,7 @@ session@1.1.7
 sha@1.0.9
 shell-server@0.2.4
 spacebars@1.0.15
-spacebars-compiler@1.1.2
+spacebars-compiler@1.1.3
 srp@1.0.10
 templating@1.3.2
 templating-compiler@1.3.2
@@ -163,5 +165,5 @@ ui@1.0.13
 underscore@1.0.10
 url@1.1.0
 vsivsi:job-collection@1.4.0
-webapp@1.3.17
+webapp@1.3.18
 webapp-hashing@1.0.9
