+++ imports/plugins/included/payments-stripe/server/methods/stripe.js
@@ -4,6 +4,7 @@ import accounting from "accounting-js";
 import { Meteor } from "meteor/meteor";
 import { check, Match } from "meteor/check";
 // reaction modules
+import { Packages } from "/lib/collections";
 import { Reaction, Logger } from "/server/api";
 import { StripeApi } from "./stripeapi";
 
@@ -46,8 +47,20 @@ function parseCardData(data) {
 function formatForStripe(amount) {
   return Math.round(amount * 100);
 }
+
 function unformatFromStripe(amount) {
   return (amount / 100);
+}
+
+export function getApiKey() {
+  const settings = Packages.findOne({
+    name: "reaction-stripe",
+    shopId: Reaction.getShopId()
+  }).settings;
+  if (!settings.api_key) {
+    throw new Meteor.Error("403", "Invalid Stripe Credentials");
+  }
+  return settings.api_key;
 }
 
 function stripeCaptureCharge(paymentMethod) {
@@ -57,9 +70,11 @@ function stripeCaptureCharge(paymentMethod) {
   };
 
   try {
+    const apiKey = getApiKey();
     const captureResult = StripeApi.methods.captureCharge.call({
       transactionId: paymentMethod.transactionId,
-      captureDetails: captureDetails
+      captureDetails,
+      apiKey
     });
     if (captureResult.status === "succeeded") {
       result = {
@@ -117,7 +132,8 @@ Meteor.methods({
     let chargeResult;
 
     try {
-      chargeResult = StripeApi.methods.createCharge.call({ chargeObj });
+      const apiKey = getApiKey();
+      chargeResult = StripeApi.methods.createCharge.call({ chargeObj, apiKey });
       if (chargeResult && chargeResult.status && chargeResult.status === "succeeded") {
         result = {
           saved: true,
@@ -182,7 +198,8 @@ Meteor.methods({
 
     let result;
     try {
-      const refundResult = StripeApi.methods.createRefund.call({ refundDetails });
+      const apiKey = getApiKey();
+      const refundResult = StripeApi.methods.createRefund.call({ refundDetails, apiKey });
       Logger.debug(refundResult);
       if (refundResult && refundResult.object === "refund") {
         result = {
@@ -216,7 +233,8 @@ Meteor.methods({
     check(paymentMethod, Reaction.Schemas.PaymentMethod);
     let result;
     try {
-      const refunds = StripeApi.methods.listRefunds.call({ transactionId: paymentMethod.transactionId });
+      const apiKey = getApiKey();
+      const refunds = StripeApi.methods.listRefunds.call({ transactionId: paymentMethod.transactionId, apiKey });
       result = [];
       for (const refund of refunds.data) {
         result.push({
